moved react components into seperate folder
