+++ src/ChatroomComponents/Chatroom.js
@@ -4,7 +4,7 @@ import Footer from './Footer';
 import Sidebar from './Sidebar';
 import MessageBoard from './MessageBoard';
 import RoomList from './RoomList';
-import { listenForMessages } from './client.js';
+import { listenForMessages } from '../client.js';
 
 class Chatroom extends Component {
   constructor(props) {
