+++ src/ChatroomComponents/MessageBoard.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import Message from './Message';
-import { updateScroll } from './helpers';
+import { updateScroll } from '../helpers';
 
 class MessageBoard extends Component{
